+++ exercise.main/Product.cs
@@ -9,11 +9,11 @@ namespace exercise.main
     public class Product
     {
         private string _sku;
-        private double _price;
+        private decimal _price;
         private ProductType _type;
         private string _variant;
 
-        public Product(string sku, double price, ProductType type, string variant)
+        public Product(string sku, decimal price, ProductType type, string variant)
         {
             _sku = sku;
             _price = price;
@@ -22,7 +22,7 @@ namespace exercise.main
         }
 
         public string SKU { get { return _sku;} set { _sku = value; } }
-        public double Price { get { return _price;} set { _price = value; } }
+        public decimal Price { get { return _price;} set { _price = value; } }
         public ProductType Type { get { return _type; } set { _type = value; } }
         public string Variant { get { return _variant; } set { _variant = value; } }
